requirements.txt
allure-pytest==2.8.5
allure-python-commons==2.8.5
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
bcrypt==3.1.7
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
idna==2.8
importlib-metadata==0.23
more-itertools==7.2.0
packaging==19.2
paramiko==2.6.0
pluggy==0.13.0
psycopg2==2.8.3
py==1.8.0
pycparser==2.19
PyNaCl==1.3.0
pyparsing==2.4.2
pytest==5.1.2
requests==2.22.0
six==1.12.0
urllib3==1.25.3
wcwidth==0.1.7
zipp==0.6.0
